[c] OUT-44 - Refactor hash authenticate code
