+++ supabase/functions/user-actions/index.ts
@@ -3,7 +3,6 @@ import {handleError, replacePlaceholders} from "./utils.ts";
 import { handleNewComment } from "./handlers/comment-handler.ts";
 import { handleTeamChange } from "./handlers/team-handler.ts";
 import { markdownTemplateHeader, markdownTemplateBody } from "./templates/slack.ts"
-import { decodeHex } from "https://deno.land/std@0.210.0/encoding/hex.ts";
 import { SlackAPI } from "https://deno.land/x/deno_slack_api@2.1.1/mod.ts";
 
 import { drizzle, PostgresJsDatabase } from "npm:drizzle-orm/postgres-js";
@@ -14,6 +13,7 @@ import {
   handleConversationClosed,
 } from "./handlers/conversation-handler.ts";
 import { handleTwilioMessage } from "./handlers/twilio-message-handler.ts";
+import {Verify} from "./authenticate";
 
 const client = postgres(Deno.env.get("DB_POOL_URL")!, { prepare: false });
 const db: PostgresJsDatabase = drizzle(client);
@@ -48,36 +48,15 @@ Deno.serve(async (req) => {
     await sendToSlack(client, replacementDictionary);
     return new Response("", { status: 400 });
   } else {
-    const encoder = new TextEncoder();
-    const data = encoder.encode(JSON.stringify(requestBody));
-
-    const keyBuf = encoder.encode(Deno.env.get("HMAC_SECRET")!);
-
-    const key = await crypto.subtle.importKey(
-        "raw",
-        keyBuf,
-        { name: "HMAC", hash: "SHA-256" },
-        true,
-        ["sign", "verify"],
-    );
-
-    const keyPrefix = "sha256=";
-      const cleanedHeaderSig = requestHeaderSig.startsWith(keyPrefix)
-          ? requestHeaderSig.slice(keyPrefix.length)
-          : requestHeaderSig;
-
-      const receivedSignature = decodeHex(cleanedHeaderSig);
-
-      const verified = await crypto.subtle.verify({ name: "HMAC", hash: "SHA-256" }, key, receivedSignature, data.buffer);
-
-      if (!verified){
-        replacementDictionary.failureHost = clientIps;
-        replacementDictionary.failureDetails = "Mismatched authentication header";
-        replacementDictionary.failedRequestDetails = JSON.stringify(requestBody, null, 2).toString();
-        replacementDictionary.failedRule = requestBody.rule.type;
-        await sendToSlack(client, replacementDictionary);
-        return new Response("", { status: 400 });
-      }
+    const verified = await Verify(requestHeaderSig, requestBody);
+    if (!verified){
+      replacementDictionary.failureHost = clientIps;
+      replacementDictionary.failureDetails = "Mismatched authentication header";
+      replacementDictionary.failedRequestDetails = JSON.stringify(requestBody, null, 2).toString();
+      replacementDictionary.failedRule = requestBody.rule.type;
+      await sendToSlack(client, replacementDictionary);
+      return new Response("", { status: 400 });
+    }
   }
 
   try {
